金额修改为浮点

+++ admin/models.py
@@ -81,7 +81,7 @@ class CustomerCard(db.Model):
     card = db.relationship(Card)
     card_code = db.Column(db.String(32))  # 动态分配的卡号
     img = db.Column(db.String(36))
-    amount = db.Column(db.Integer())
+    amount = db.Column(db.Float())
     claimed_time = db.Column(db.DateTime(), default=datetime.now)
     wx_binding_time = db.Column(db.DateTime())
     expire_date = db.Column(db.Date())
